build(deps): bump buildx version to v0.23.0
Github cache API v1 has been sunset and the usage of that cache API
by docker/build-push-action is causing all builds to fail. Support
for newer (and still supported) versions of the cache API is possible
in newer versions of buildx, so we are bumping to latest here. More
details can be found in https://docs.docker.com/build/ci/github-actions/cache/#github-cache.

Signed-off-by: Jared Watts <jbw976@gmail.com>